clipboard-apis/readText.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Clipboard API: navigator.clipboard.readText()</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
  <h1>Test for navigator.clipboard.readText()</h1>

  <script>
    promise_test(async t => {
      try {
        const text = await navigator.clipboard.readText();
        assert_equals(typeof text, "string", "readText() should resolve to a string");
      } catch (e) {
        assert_unreached("Clipboard readText() threw an error: " + e.message);
      }
    }, "navigator.clipboard.readText() resolves with a string");
  </script>
</body>
